refactor: 설정 화면 파일 구조 분리
