+++ presentation/src/main/java/com/team/yeogibeoryeo/presentation/settings/SettingsLayoutDefaults.kt
@@ -1,0 +1,16 @@
+package com.team.yeogibeoryeo.presentation.settings
+
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+
+internal object SettingsLayoutDefaults {
+    val listHorizontalPadding: Dp = 24.dp
+    val listVerticalPadding: Dp = 16.dp
+    val listItemVerticalPadding: Dp = 24.dp
+    val detailHorizontalPadding: Dp = 24.dp
+    val detailVerticalPadding: Dp = 24.dp
+    val detailItemSpacing: Dp = 24.dp
+    val detailContentSpacing: Dp = 20.dp
+    val sectionSpacing: Dp = 8.dp
+    val infoRowSpacing: Dp = 8.dp
+}
